refactor: screen을 calculator파일에서 분리
